Add files via upload
This is the Jupyter notebook .ipynb file which is used for training our LSTM model for the stock price prediction of Google. It contains extensive data-exploration, for better understanding and visualization of the dataset, as well as data pre-processing for handling Nan values and creating new features from the already present features in the dataset.

To train the LSTM model, we first find the best hyper-parameters and perform Bayesian optimization. The model is trained on the best hyper-parameters so that the best results can be achieved. The visualization of the predicted and actual model is done for a better understanding of the models performance. 

Next, for predicting on live stock market rate, we use the Alpha Vantage API key and load the data into our .ipynb notebook. The same pre-processing is done as during the training of the model, before predicting the outputs on the live data.